refactp verso test getters
